rename: featuers/player index 관리(#206)
