+++ src/features/game/components/ranking/common/TeamRankingView.tsx
@@ -1,7 +1,7 @@
 import { CustomBarChart, DataTable } from '@/features/common';
 import { TeamBatterRank } from '@/features/common/types/batters';
 import { TeamPitcherRank } from '@/features/common/types/pitchers';
-import { Config } from '@/features/player/components';
+import { Config } from '@/features/player';
 import { RecentRecord, YearRecord } from '@/features/player/types/detail';
 import { ColumnDef } from '@tanstack/react-table';
 import { useState } from 'react';
