+++ src/features/player/components/detail/profile/LeagueRecord.tsx
@@ -89,4 +89,4 @@ function LeagueRecord() {
   );
 }
 
-export default LeagueRecord;
+export { LeagueRecord };
