+++ src/features/player/components/detail/profile/PlayerInfo.tsx
@@ -62,4 +62,4 @@ function PlayerInfo() {
   );
 }
 
-export default PlayerInfo;
+export { PlayerInfo };
