--- src/features/player/components/index.ts
@@ -1,10 +1,0 @@
-export { PlayerList } from './PlayerList';
-export { PlayerProfile } from './detail/PlayerProfile';
-export { PlayerRecordChart } from './detail/PlayerRecordChart';
-export type { Config } from './detail/PlayerRecordChart';
-export { RecordTableAccordion } from './detail/RecordTableAccordion';
-export { SeasonSummary } from './detail/SeasonSummary';
-export { StatCard } from './detail/StatCard';
-
-/* Skeletons */
-export { PlayerProfileSkeleton } from './skeletons/PlayerProfileSkeleton';
