+++ src/pages/player/BatterPage.tsx
@@ -5,8 +5,7 @@ import {
   SearchBar,
   SubTabsTrigger,
 } from '@/features/common';
-import { PlayerList } from '@/features/player/components';
-import NotFoundSearch from '@/features/player/components/NotFoundSearch';
+import { NotFoundSearch, PlayerList } from '@/features/player';
 import { usePlayerSearch } from '@/features/player/hooks/usePlayerSearch';
 import { useTabFromUrl } from '@/hooks/useTabFromUrl';
